v1.1.2
Stability improvement for the underwater fix: the NiPoint3 I was using for the camera more often correlates to its 3D node's local position, when it's the world position we need to use.

Added a fix for loading the ARMA/MO5S subrecord. First-person texture swaps still don't appear to work in-game (for any gender), but this allows the swap data to load into the TESObjectARMA::models[1][1] properly, at least.