INI handling
The merchant restock fix is now placed behind an INI setting.

Replaced our INI loading/saving code with code from NorthernUI, with the added change of supporting unsigned-int settings. This code will automatically load all INI settings en masse (instead of me manually having to load them), and when saving an INI file, it will attempt to keep settings in the same order (and with the same comments) as the user's original file (this is less relevant for CobbBugFixes; for NorthernUI, we wanted it for changing settings from in-game).