v1.0.4
Added a fix for a Skyrim engine bug that would often, but inconsistently, prevent torches held by NPCs from lighting the landscape.

Added version information to the DLL (visible when viewing the file properties in Windows).

Added a crash logger, which is intended to dump CPU state at the time of the crash.