v1.6.1
The project now builds in VS2019.

The crash logger has been improved: modules are now sorted by memory address before being listed; if registers' values line up with module address ranges, then the modules and module-relative offsets are shown; if values on the stack are pointers to identifiable objects, then the mangled class names are shown; if values on the stack line up with module address ranges, then the modules and module-relative offsets are shown; and the number of stack values shown can be configured using an INI setting.